requirements.txt
# IEC 61131-3 Code Generator - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
pymongo>=4.6.0

# AI/ML
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Data Validation
pydantic>=2.5.0
jsonschema>=4.20.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Type hints (for development)
typing-extensions>=4.8.0
